added serializer exception, changed fmt type, added settings
